+++ b_cfn_lambda_layer/tmp/__init__.py
@@ -1,0 +1,3 @@
+from os.path import dirname, abspath
+
+docker_build_root = dirname(abspath(__file__))