+++ b_cfn_lambda_layer_test/integration/infrastructure/function3.py
@@ -3,32 +3,32 @@ from aws_cdk.core import Stack
 from b_aws_testing_framework.tools.cdk_testing.testing_stack import TestingStack
 
 from b_cfn_lambda_layer.lambda_layer import LambdaLayer
-from b_cfn_lambda_layer_test.integration.infrastructure.layer_source_2 import root
+from b_cfn_lambda_layer_test.integration.infrastructure.layer1 import root as root1
+from b_cfn_lambda_layer_test.integration.infrastructure.layer2 import root as root2
+from b_cfn_lambda_layer_test.integration.infrastructure.layer3 import root as root3
 
 
 class Function3(Function):
     """
-    Function that allows us to test whether installing dependencies
-    via requirements.txt works.
+    Function that allows us to test whether multiple same layers override each other.
     """
     def __init__(self, scope: Stack):
         super().__init__(
             scope=scope,
             id=f'{TestingStack.global_prefix()}TestingFunction3',
             code=Code.from_inline(
-                'import urllib3\n'
-                # Jose comes from requirements.txt.
-                'from jose import jwk, jwt\n'
-                'from jose.utils import base64url_decode\n'
-                'import boto3\n'
-                'import botocore\n'
-                # Faker comes from requirements.txt.
-                'from faker.utils import text\n'
+                # Ensure that dummy module is accessible from lambda layer
+                # and the parent directory is included.
+                'from layer1.dummy_module import DummyModule as M1\n'
+                'from layer2.dummy_module import DummyModule as M2\n'
+                'from layer3.dummy_module import DummyModule as M3\n'
                 '\n\n'
                 'def handler(*args, **kwargs):\n'
                 '    return dict(\n'
-                '        Boto3Version=boto3.__version__,\n'
-                '        BotocoreVersion=botocore.__version__,\n'
+                # Use the dummy module to ensure everything works.
+                '        Dummy1=M1.action(),\n'
+                '        Dummy2=M2.action(),\n'
+                '        Dummy3=M3.action()\n'
                 '    )'
                 '\n'
             ),
@@ -37,10 +37,29 @@ class Function3(Function):
             layers=[
                 LambdaLayer(
                     scope=scope,
-                    name=f'{TestingStack.global_prefix()}TestingLayer3',
-                    source_path=root,
+                    layer_name=f'{TestingStack.global_prefix()}TestingLayer3_1',
+                    source_path=root1,
                     code_runtimes=[Runtime.PYTHON_3_6, Runtime.PYTHON_3_7, Runtime.PYTHON_3_8],
-                    include_source_path_directory=False
+                ),
+                LambdaLayer(
+                    scope=scope,
+                    layer_name=f'{TestingStack.global_prefix()}TestingLayer3_2',
+                    source_path=root2,
+                    code_runtimes=[Runtime.PYTHON_3_6, Runtime.PYTHON_3_7, Runtime.PYTHON_3_8],
+                ),
+
+                # Repeat same layer twice.
+                LambdaLayer(
+                    scope=scope,
+                    layer_name=f'{TestingStack.global_prefix()}TestingLayer3_3_1',
+                    source_path=root3,
+                    code_runtimes=[Runtime.PYTHON_3_6, Runtime.PYTHON_3_7, Runtime.PYTHON_3_8],
+                ),
+                LambdaLayer(
+                    scope=scope,
+                    layer_name=f'{TestingStack.global_prefix()}TestingLayer3_3_2',
+                    source_path=root3,
+                    code_runtimes=[Runtime.PYTHON_3_6, Runtime.PYTHON_3_7, Runtime.PYTHON_3_8],
                 )
             ]
         )