+++ b_cfn_lambda_layer_test/integration/tests/test_lambda_layer_multiple.py
@@ -6,7 +6,7 @@ from botocore.response import StreamingBody
 from b_cfn_lambda_layer_test.integration.infrastructure.main_stack import MainStack
 
 
-def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_layer_with_parent_dir_EXPECT_execution_successful():
+def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_EXPECT_execution_successful():
     """
     Test whether the layer provides necessary functionality.
 
@@ -17,7 +17,7 @@ def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_layer_with_parent_d
 
     # Invoke specific lambda function.
     response = lambda_client.invoke(
-        FunctionName=MainStack.get_output(MainStack.LAMBDA_FUNCTION_4_NAME_KEY),
+        FunctionName=MainStack.get_output(MainStack.LAMBDA_FUNCTION_3_NAME_KEY),
         InvocationType='RequestResponse'
     )
 
@@ -27,4 +27,6 @@ def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_layer_with_parent_d
     data = json.loads(''.join(data))
 
     # Assert that the result is as expected.
-    assert data.get('Dummy') == 'Hello world from dummy module!', data
+    assert data.get('Dummy1') == 'Layer1', data
+    assert data.get('Dummy2') == 'Layer2', data
+    assert data.get('Dummy3') == 'Layer3', data